requirements.txt
asgiref==3.11.1
diff-match-patch==20241021
Django==6.0.2
django-cors-headers==4.1.0
django-filter==25.2
django-import-export==4.4.0
django-jazzmin==2.6.0
django-multiselectfield==1.0.1
djangorestframework==3.14.0
gunicorn==24.1.1
mysqlclient==2.2.7
packaging==26.0
pillow==10.4.0
pytz==2024.1
sqlparse==0.5.1
tablib==3.9.0
typing_extensions==4.12.2
